requirements.txt
requests==2.32.3
google-generativeai==0.8.5
garminconnect==0.2.30
python-dotenv==1.0.1
